// src/app/page/master/pegawai-struktur-gaji/pegawai-struktur-gaji.interface.ts
import { EventEmitter } from '@angular/core';

export interface  PegawaiStrukturGaji {
    kode?: any;
    kdGolonganPegawai?: number;
    hargaSatuan?: string
    kdOperatorRate?: string;
    kdKomponenHarga?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}